Add docstrings, str stream combination

+++ build.bash
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-rm -rf artifacts/
+if [[ -n $1 ]];
+then
+    rm -rf artifacts/
+fi
 
 cmake -S . -B artifacts/
 cmake --build artifacts --parallel
